json <-> cgi
